02_geneious_analysis Overhaul
The "test.geneiousWorklfow" has been removed and has been replaced by the  "MaRS_Geneious_workflow_V1.geneiousWorkflow" workflow that Dhruvi Patel developed and that Marko Bajic tested several times and had collaborators test at least twice in Senegal and can confirm works.

Along with the revamp of this workflow, the associated file "MaRS_ReferenceGenes.geneious" has also been added to the folder (it's small in size to be shared, and it belongs in this folder!). This is the same content as the "Updated_MaRS_refs_ET.geneious" file but the name has been changed to be more logical, clear, and also the names of genes has been changed to follow the Pf naming scheme (Pfk13 instead of k13, for example).

As this naming change affects the VOI information, the "voinew3.csv" file has also been updated. This naming change would also affect the report generation but that is not currently covered on Github. There is a mention of the genes in "DF_Analysis1_Angola2.ipynb" in 03_summary_tables > Epi report, but only in the descriptive section, the name change would not actually affect the jupyter code as it is currently written.

The MOST noteworthy changes are in the 02_geneious_analysis\README.md. The file is much more comprehensive now and is outlined in similar way, and utilizing the same flow and powerful organizational structure as the Lab SOP uses. The file now also references the new, currently used Workflow and the most up-to-date reference file, and it also covers in more detail how to import both the workflow and reference file into Geneious.