Massive Overhaul of MaRS
All the changes regarding the readme files are (I tried to the best of my abilities) outlined at the top/beginning of the readme file with a dropdown that shows what has been done and what still needs to be done.

Removed the no longer used images we had in /images/ and /docs/assets/

Created two new images: one demonstrates how the GitHub is organized (folder structure), and the other outlines the MaRS workflow, with emphasis on the Geneious steps.

Added two exmple Excel files to 04_ncbi_submission to help with demonstrating how to fill out two essential files required for NCBI submissions.

Outlined and started to combine 03_ and old 04_data_vis into one new folder called 03_geneious_reports which will contain a single readme file, VariantsOfInterest.csv file, and a Jupyter notebook file where the two outputs are created using the single output from Geneious, as well as doing visualization of the results. The three Epi/Lab report files are still in there, but the plan is to remove these in the future as the one single .ipynb file is finished

Added information about Bowtie2 and how to overcome a strange new error that was discovered where, for whatever reason, the newest version will not produce alignment results for some files. It was originally though this was because of how it was packaged within our workflow but even mapping by itself still does not work. This issue does not exist with the former version of Bowtie2 and that is why the plugin is also included in the directory. Instructions on how to install this specific directory are outlined in the readme file. Also added SNPs_and_Haplotypes for control strains excel file that can be used for evaluating results.