Workflow Update - One Option Fixes the BBDuk Issue
The notorious "If BBDuk creates an output file with no reads it will kill the workflow" error is finally overcome. For most of the steps in the workflow there is a sneaky (you have to expand the window to see it) option at the very bottom of the bottom that says "If this workflow step fails, silently continue using only successful data sets" and it overcomes the issue we had. Our issue was that Geneious prioritized reporting the error over running the workflow, which is unfortunate.

Confirmed, the new change in the Workflow makes it work perfectly, even if a nasty (nothing but bad quality reads) file is present. Briliant!

Now that this is fixed, the blurb about this has been removed from the 02_geneious_analysis/readme file